chore(tools): relocate player extractor
